+++ src/components/CountrySelector/Options.tsx
@@ -1,22 +1,25 @@
 import { Combobox } from "@headlessui/react";
 import React from "react";
-import { Country } from "types/countries";
-import countries from "assets/countries/all.json";
+import { Country } from "types/components";
 
 const containerStyle =
   "absolute top-full mt-2 z-10 w-full bg-white dark:bg-blue-d6 shadow-lg rounded overflow-y-scroll scroller";
 
-const options: Country[] = countries;
-export default function Options({ query }: { query: string }) {
+type Props = {
+  query: string;
+  options: Country[];
+};
+
+export default function Options({ query, options }: Props) {
   const filtered = options.filter((o) =>
     o.name.toLocaleLowerCase().includes(query.toLocaleLowerCase())
   );
   return (
     <Combobox.Options
       className={containerStyle}
-      style={{ height: countries.length <= 10 ? "auto" : "10rem" }}
+      style={{ height: options.length <= 10 ? "auto" : "10rem" }}
     >
-      {(countries.length > 0 &&
+      {(options.length > 0 &&
         filtered.map((country) => (
           <Combobox.Option
             as={React.Fragment}
