+++ src/components/CountrySelector/constants.ts
@@ -1,0 +1,7 @@
+import { Country } from "types/components";
+
+export const placeHolderCountryOption: Country = {
+  name: "",
+  flag: "",
+  code: "",
+};
