+++ src/components/FileDropzone/helpers.ts
@@ -1,0 +1,7 @@
+import { FileObject } from "types/aws";
+import { Asset } from "types/components";
+
+export const asset = (previews: FileObject[]): Asset => ({
+  files: [],
+  previews,
+});
