+++ src/components/FileDropzone/index.ts
@@ -1,0 +1,3 @@
+export { default } from "./FileDropzone";
+export type { Asset } from "types/components";
+export { asset } from "./helpers";
