+++ src/components/Selector/Selector.tsx
@@ -2,7 +2,7 @@ import { Listbox } from "@headlessui/react";
 import { ErrorMessage } from "@hookform/error-message";
 import { FieldValues, Path, get, useController } from "react-hook-form";
 import { Props } from "./types";
-import { OptionType, ValKey } from "types/utils";
+import { OptionType, ValKey } from "types/components";
 import { DrawerIcon } from "components/Icon";
 import FocusableInput from "./FocusableInput";
 import { styles, valueKey } from "./constants";
