+++ src/components/registration/index.ts
@@ -1,3 +1,2 @@
-export * from "./FileDropzone";
 export * from "./LoadText";
 export * from "./Separator";
