+++ src/pages/Admin/Charity/EditProfile/getSDGLabelValuePair.ts
@@ -1,5 +1,5 @@
+import { OptionType } from "types/components";
 import { UNSDG_NUMS } from "types/lists";
-import { OptionType } from "types/utils";
 
 export const getSDGLabelValuePair = (
   key: string | number,
