+++ src/pages/Admin/Charity/EditProfile/types.ts
@@ -1,8 +1,8 @@
 import { OverrideProperties } from "type-fest";
 import { EndowDesignation, EndowmentProfileUpdate } from "types/aws";
-import { Country } from "types/countries";
+import { OptionType } from "types/components";
+import { Country } from "types/components";
 import { UNSDG_NUMS } from "types/lists";
-import { OptionType } from "types/utils";
 import { ImgLink } from "components/ImgEditor";
 
 export type FV = OverrideProperties<
