+++ src/pages/Registration/Steps/ContactDetails/index.tsx
@@ -2,7 +2,7 @@ import { yupResolver } from "@hookform/resolvers/yup";
 import { FormProvider, useForm } from "react-hook-form";
 import { FormValues } from "./types";
 import { ContactRoles, ReferralMethods } from "types/aws";
-import { OptionType } from "types/utils";
+import { OptionType } from "types/components";
 import { useRegState, withStepGuard } from "../StepGuard";
 import Form from "./Form";
 import { referralMethods, roles } from "./constants";
