+++ src/pages/Registration/Steps/Documentation/Form/index.tsx
@@ -5,9 +5,10 @@ import countries from "assets/countries/all.json";
 import ActivityCountries from "components/ActivityCountries";
 import CountrySelector from "components/CountrySelector";
 import ExtLink from "components/ExtLink";
+import FileDropzone from "components/FileDropzone";
 import { MultiSelector, Selector } from "components/Selector";
 import { Field, Label, Radio } from "components/form";
-import { FileDropzone, LoadText } from "components/registration";
+import { LoadText } from "components/registration";
 import { MB_LIMIT } from "schemas/file";
 import { unsdgs } from "constants/unsdgs";
 import { TERMS_OF_USE_NPO } from "constants/urls";
