+++ src/pages/Registration/Steps/Documentation/index.tsx
@@ -1,8 +1,8 @@
 import { yupResolver } from "@hookform/resolvers/yup";
 import { FormProvider, useForm } from "react-hook-form";
 import { FormValues } from "./types";
+import { asset } from "components/FileDropzone";
 import { useRegState, withStepGuard } from "../StepGuard";
-import { genFileAsset } from "../getRegistrationState";
 import Form from "./Form";
 import { schema } from "./schema";
 
@@ -19,8 +19,8 @@ function Documentation() {
         }
       : {
           ein: "",
-          proofOfIdentity: genFileAsset([]),
-          proofOfRegistration: genFileAsset([]),
+          proofOfIdentity: asset([]),
+          proofOfRegistration: asset([]),
           website: "",
 
           sdgs: [],
