+++ src/pages/Registration/Steps/getRegistrationState.ts
@@ -8,12 +8,11 @@ import {
   ContactDetails,
   DoneContact,
   DoneDocs,
-  FileObject,
   InitContact,
   SavedRegistration,
   TDocumentation,
 } from "types/aws";
-import { Asset } from "components/registration";
+import { asset } from "components/FileDropzone";
 import { unsdgs } from "constants/unsdgs";
 
 export function getRegistrationState(
@@ -95,8 +94,8 @@ function formatDocumentation({
 }: DoneDocs["Registration"]): Documentation {
   return {
     //level 1
-    proofOfIdentity: genFileAsset([poi]),
-    proofOfRegistration: genFileAsset([por]),
+    proofOfIdentity: asset([poi]),
+    proofOfRegistration: asset([por]),
     ein: EIN,
 
     website: Website,
@@ -121,10 +120,6 @@ function formatDocumentation({
     tier: Tier,
     isAnonymousDonationsAllowed: KycDonorsOnly ? "No" : "Yes",
   };
-}
-
-export function genFileAsset(previews: FileObject[]): Asset {
-  return { files: [], previews };
 }
 
 function isDoneDocs(data: SavedRegistration): data is DoneDocs {
