+++ src/schemas/shape.ts
@@ -1,7 +1,7 @@
 import { lazy, object, string } from "yup";
 import { SchemaShape } from "./types";
+import { OptionType } from "types/components";
 import type { TokenWithAmount as TWA } from "types/tx";
-import { OptionType } from "types/utils";
 import { tokenConstraint } from "./number";
 import { requiredString } from "./string";
