--- src/types/countries.ts
@@ -1,5 +1,0 @@
-export type Country = {
-  name: string;
-  flag: string;
-  code: string;
-};
